major revamp of core features
big update, lemme try to remember all that's in it.

sprites and complex sprites were completely reworked from the ground up, they are still created and drawn to the screen in the same way, so no backwards compatibility issues.
Sprites and Complex sprites now use a different method of displaying, manually editing pygame buffers instead of using set_at, which drastically increases draw times with large sprites and has marginal benefits with small sprites. the old method of sprites is still implemented but their implementations have been changed to legacySprite, legacyComplexSprite, putLegacySprite and putLegacyComplexSprite. (These might still be useful because new sprites might take up more memory, but I haven't tested that yet)

New kind of sprite has been implemented, sizeSprite. a sizeSprite is similar to a normal sprite in that it is created from a single image, but it is also initialized with a width and a height in tiles. when drawn on the screen the image will be scaled to that width and height, breaking the reliance on simple 1x1 size sprites. These sprites can also be initialized with decimal width, ie 1.75 tiles wide and 4.4 tiles height. 

that should cover the sprites now onto some other things. Im not sure if this was in an update before but its here to stay now, screenshots. simply call graphicsObject.screenshot(<imagename>) to save an image of the current screen

also, automatic window scaling is not implemented into tileGraphics, now every screen you initialize with tileGraphics can be scaled manually and tileGraphics will automatically scale the screen to the new resolution, this includes pressing F11 to automatically toggle between fullscreen and normal resolution. while this scaling is not perfect as tilegraphics simply scales the image as is, in hopes that it will look decent enough, subsequent draws to the screen will normally get the screen looking decent in a short enough timespan. new sprites are also automatically reloaded if the resolution has changed in a way which will impact their rendering this reloading happens once a sprite is called to be rendered so old unused sprites will not be reloaded until they are used again. 

 the screen can also now be resized to a new number of tiles by using the graphicsObject.resize(<width>,<height>) command

I think that's it.